feat : 게시물 삭제시 caseCade ALL 모두 다 삭제 구현

+++ backend/src/main/java/com/newcen/newcen/question/service/QuestionService.java
@@ -115,7 +115,7 @@ public class QuestionService {
         BoardEntity boardGet = questionsRepositorySupport.findBoardByUserIdAndBoardId(userId,boardId);
         UserEntity user = userRepository.findById(userId).get();
         List<BoardFileEntity> boardFileEntityList = boardFileRepository.findByBoardId(boardId);
-        if (boardFileEntityList.size() !=0 ){
+        if (boardFileEntityList.size() !=0 && boardFileEntityList !=null){
             boardFileEntityList.forEach(t-> amazonS3.deleteObject(new DeleteObjectRequest(bucket, t.getBoardFilePath())));
         }
         if (!Objects.equals(boardGet.getUserId(), user.getUserId())){
